+++ scripts/secrets-help.sh
@@ -24,10 +24,12 @@ Common commands:
   make secrets-help
   make account
   make password
+  make rotate-password
   make secret-check
   make keychain-add ACCOUNT_NAME=work
   make password-file-init ACCOUNT_NAME=work
   make gpg-file-init ACCOUNT_NAME=work GPG_RECIPIENT='<your key id>'
+  make rotate-password ACCOUNT_NAME=work
 
 Accounts directory usage: